chore: remove debug utils file
